phase 1 passed

+++ src/main/java/org/yearup/service/ProductService.java
@@ -1,6 +1,8 @@
 package org.yearup.service;
 
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+import org.yearup.exception.ResourceNotFoundException;
 import org.yearup.models.Product;
 import org.yearup.repository.ProductRepository;
 
@@ -10,7 +12,7 @@ import java.util.List;
 public class ProductService
 {
     private final ProductRepository productRepository;
-
+    @Autowired
     public ProductService(ProductRepository productRepository)
     {
         this.productRepository = productRepository;
@@ -30,14 +32,14 @@ public class ProductService
                        .toList();
     }
 
-    public List<Product> listByCategoryId(int categoryId)
+    public List<Product> listProductsByCategoryId(int categoryId)
     {
-        return productRepository.findByCategoryId(categoryId);
+        return productRepository.findByCategoryId(categoryId); // add error if category id doesnt exist.
     }
 
     public Product getById(int productId)
     {
-        return productRepository.findById(productId).orElse(null);
+        return productRepository.findById(productId).orElseThrow(()-> new ResourceNotFoundException("Product Not Found: " + productId));
     }
 
     public Product create(Product product)
@@ -48,19 +50,23 @@ public class ProductService
 
     public Product update(int productId, Product product)
     {
-        Product existing = productRepository.findById(productId).orElseThrow();
-        existing.setName(product.getName());
-        existing.setPrice(product.getPrice());
-        existing.setCategoryId(product.getCategoryId());
-        existing.setDescription(product.getDescription());
-        existing.setSubCategory(product.getSubCategory());
-        existing.setFeatured(product.isFeatured());
-        existing.setImageUrl(product.getImageUrl());
-        return productRepository.save(existing);
+        return productRepository.findById(productId).map(existing -> {
+            existing.setName(product.getName());
+            existing.setPrice(product.getPrice());
+            existing.setCategoryId(product.getCategoryId());
+            existing.setDescription(product.getDescription());
+            existing.setSubCategory(product.getSubCategory());
+            existing.setFeatured(product.isFeatured());
+            existing.setImageUrl(product.getImageUrl());
+            return productRepository.save(existing);
+        }).orElseThrow(()-> new ResourceNotFoundException("Product Not Found: " + productId));
     }
 
     public void delete(int productId)
     {
+        if (!productRepository.existsById(productId)){
+            throw new ResourceNotFoundException("Product Not Found: " + productId);
+        }
         productRepository.deleteById(productId);
     }
 }
